[test.la_container] split into three tests and use dune-testtools
