+++ dune/stuff/test/la_container.mini
@@ -1,0 +1,12 @@
+__name = la_container
+__exec_suffix = {container}_{fieldtype_short}
+
+include vectors.mini
+
+include matrices.mini
+
+container = {vector}, {matrix} | expand
+
+[__static]
+TESTCONTAINERTYPE = Dune::Stuff::LA::{container}<fieldtype>
+
