+++ dune/stuff/test/la_container_vector.mini
@@ -1,0 +1,7 @@
+__name = la_container_vector
+__exec_suffix = {vector}_{fieldtype_short}
+
+include vectors.mini
+
+[__static]
+TESTVECTORTYPE = Dune::Stuff::LA::{vector}<fieldtype>
