+++ dune/stuff/test/matrices.mini
@@ -1,0 +1,6 @@
+matrix_common = CommonDenseMatrix
+matrix_eigen = EigenRowMajorSparseMatrix, EigenDenseMatrix | expand
+matrix_istl = IstlRowMajorSparseMatrix
+matrix = {matrix_common}, {matrix_eigen}, {matrix_istl} | expand types
+
+1, EIGEN_FOUND, dune-istl_FOUND | expand types | cmake_guard
