+++ dune/stuff/test/vectors.mini
@@ -1,0 +1,9 @@
+fieldtype = double, std::complex<double> | expand field
+fieldtype_short = double, complex | expand field
+
+vector_common = CommonDenseVector
+vector_eigen = EigenDenseVector, EigenMappedDenseVector | expand
+vector_istl = IstlDenseVector
+vector = {vector_common}, {vector_eigen}, {vector_istl} | expand types
+
+1, EIGEN_FOUND, dune-istl_FOUND | expand types | cmake_guard
